FE Siswa, Admin, OP

+++ database/migrations/2025_01_27_154916_create_orang_tua_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('orang_tua', function (Blueprint $table) {
             $table->increments('id_orang_tua');
-            $table->unsignedInteger('id_calon_siswa');
+            $table->unsignedBigInteger('id_calon_siswa');
             $table->unsignedInteger('id_hubungan');
             $table->string('nama_lengkap', 100);
             $table->unsignedInteger('pekerjaan');
